Get rid of conftest file in tables plugin tests
