--- sardes/plugins/tables/tests/conftest.py
@@ -1,57 +1,0 @@
-# -*- coding: utf-8 -*-
-# -----------------------------------------------------------------------------
-# Copyright © SARDES Project Contributors
-# https://github.com/cgq-qgc/sardes
-#
-# This file is part of SARDES.
-# Licensed under the terms of the GNU General Public License.
-# -----------------------------------------------------------------------------
-
-# ---- Standard imports
-import os
-import os.path as osp
-from unittest.mock import Mock
-os.environ['SARDES_PYTEST'] = 'True'
-
-# ---- Third party imports
-import pytest
-
-# ---- Local imports
-from sardes.plugins.tables import SARDES_PLUGIN_CLASS
-from sardes.database.accessors import DatabaseAccessorSardesLite
-from sardes.app.mainwindow import MainWindowBase
-from sardes.database.accessors.tests.conftest import *
-
-
-@pytest.fixture
-def dbaccessor(tmp_path, database_filler):
-    dbaccessor = DatabaseAccessorSardesLite(
-        osp.join(tmp_path, 'sqlite_database_test.db'))
-    dbaccessor.init_database()
-    database_filler(dbaccessor)
-
-    return dbaccessor
-
-
-@pytest.fixture
-def mainwindow(qtbot, dbaccessor):
-    class MainWindowMock(MainWindowBase):
-        def __init__(self):
-            self.view_timeseries_data = Mock()
-            super().__init__()
-
-        def setup_internal_plugins(self):
-            self.plugin = SARDES_PLUGIN_CLASS(self)
-            self.plugin.register_plugin()
-            self.internal_plugins.append(self.plugin)
-
-    mainwindow = MainWindowMock()
-    mainwindow.show()
-    qtbot.waitExposed(mainwindow)
-
-    dbconnmanager = mainwindow.db_connection_manager
-    with qtbot.waitSignal(dbconnmanager.sig_database_connected, timeout=3000):
-        dbconnmanager.connect_to_db(dbaccessor)
-    assert dbconnmanager.is_connected()
-
-    return mainwindow
